requirements-no_torch.txt
annotated-types==0.7.0
anyio==4.9.0
beautifulsoup4==4.13.3
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
deep-translator==1.11.4
emoji==2.14.1
fastapi==0.115.12
filelock==3.18.0
fsspec==2025.3.0
h11==0.14.0
huggingface-hub==0.29.3
idna==3.10
joblib==1.4.2
nltk==3.9.1
numpy==2.2.4
packaging==24.2
pandas==2.2.3
pydantic==2.11.1
pydantic_core==2.33.0
PySastrawi @ git+https://github.com/har07/PySastrawi.git@16de9085257d9492662cd150519b852b2eb24c4c
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
safetensors==0.5.3
six==1.17.0
sniffio==1.3.1
soupsieve==2.6
starlette==0.46.1
tokenizers==0.21.1
tqdm==4.67.1
transformers==4.50.3
typing-inspection==0.4.0
typing_extensions==4.13.0
tzdata==2025.2
urllib3==2.3.0
uvicorn==0.34.0
